refactor: PublicItinerary 컴포넌트 Container 패턴으로 분리
